I think fixing the sample rate issue?
I've added some stuff in to facilitate testing on the sounds (that's
next maybe, adding unit tests...), feel free to remove if desired.

You can validate by go test - I've created a little test suite that
plays sounds with and without the fix in it. If this actually works and
it's not just a thing where my computer acts differently than yours we
can delete the flag to enable/disable the fix.